requirements.txt
python-vlc
mutagen
pyserial
